+++ R/glue.R
@@ -201,7 +201,10 @@ glue <- function(..., .sep = "", .envir = parent.frame(), .open = "{", .close = 
 
 #' Collapse a character vector
 #'
-#' Collapses a character vector of any length into a length 1 vector.
+#' `glue` collapses a character vector of any length into a length 1 vector.
+#' `glue_sql_collapse()` does the same but returns a `[DBI::SQL()]`
+#' object rather than a glue object.
+#'
 #' @param x The character vector to collapse.
 #' @param width The maximum string width before truncating with `...`.
 #' @param last String used to separate the last two items if `x` has at least